+++ src/SexyAppFramework/WidgetContainer.h
@@ -23,7 +23,7 @@ class WidgetContainer
 
 	bool mUpdateIteratorModified;
 	WidgetList::iterator mUpdateIterator;
-	ulong mLastWMUpdateCount;
+	uint32_t mLastWMUpdateCount;
 	int mUpdateCnt;
 	bool mDirty;
 	int mX;
